refactor(ai): extract CommandiumSheet composable + CommandSanitizer
