+++ manager/src/main/java/moe/shizuku/manager/logs/CommandiumSheet.kt
@@ -1,0 +1,257 @@
+@file:OptIn(
+    androidx.compose.foundation.layout.ExperimentalLayoutApi::class,
+    androidx.compose.material3.ExperimentalMaterial3Api::class,
+    androidx.compose.material3.ExperimentalMaterial3ExpressiveApi::class
+)
+
+package moe.shizuku.manager.logs
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.FlowRow
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.foundation.text.selection.SelectionContainer
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.rounded.Send
+import androidx.compose.material.icons.rounded.AutoAwesome
+import androidx.compose.material.icons.rounded.ContentCopy
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.FilterChip
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.LoadingIndicator
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.semantics.LiveRegionMode
+import androidx.compose.ui.semantics.liveRegion
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.launch
+import moe.shizuku.manager.R
+import moe.shizuku.manager.module.ModuleSettings
+import moe.shizuku.manager.utils.AiExplainUtil
+
+@Composable
+fun CommandiumSheet(
+    prompt: String,
+    onPromptChange: (String) -> Unit,
+    isGenerating: Boolean,
+    onGeneratingChange: (Boolean) -> Unit,
+    result: Result<String>?,
+    onResultChange: (Result<String>?) -> Unit,
+    scope: CoroutineScope,
+    onDismiss: () -> Unit,
+    onUseCommand: (String) -> Unit,
+    onCopy: (String) -> Unit
+) {
+    var generationJob by remember { mutableStateOf<Job?>(null) }
+    val requestCommandium: () -> Unit = {
+        if (prompt.isNotBlank() && !isGenerating) {
+            onGeneratingChange(true)
+            generationJob = scope.launch {
+                val apiKey = ModuleSettings.getComputApiKey()
+                onResultChange(AiExplainUtil.generateCommand(prompt, apiKey)))
+                onGeneratingChange(false)
+            }
+        }
+    }
+    AlertDialog(
+        onDismissRequest = {
+            onDismiss()
+            generationJob?.cancel()
+            onGeneratingChange(false)
+            onResultChange(null)
+        },
+        title = {
+            Row(
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.spacedBy(10.dp)
+            ) {
+                Icon(
+                    imageVector = Icons.Rounded.AutoAwesome,
+                    contentDescription = null,
+                    tint = MaterialTheme.colorScheme.primary,
+                    modifier = Modifier.size(24.dp)
+                )
+                Text(
+                    text = stringResource(R.string.comput_commandium_assistant),
+                    style = MaterialTheme.typography.titleLarge,
+                    fontWeight = FontWeight.Bold,
+                    color = MaterialTheme.colorScheme.onSurface
+                )
+            }
+        },
+        text = {
+            Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(20.dp)
+                .verticalScroll(rememberScrollState())
+                .imePadding(),
+            verticalArrangement = Arrangement.spacedBy(14.dp)
+        ) {
+            Text(
+                text = stringResource(R.string.comput_commandium_description),
+                style = MaterialTheme.typography.bodyMedium,
+                color = MaterialTheme.colorScheme.onSurfaceVariant
+            )
+
+            FlowRow(
+                horizontalArrangement = Arrangement.spacedBy(6.dp),
+                verticalArrangement = Arrangement.spacedBy(6.dp)
+            ) {
+                listOf(
+                    "List user installed apps",
+                    "Check battery temperature & status",
+                    "Find files larger than 50MB",
+                    "Get device Android model & build"
+                ).forEach { suggestion ->
+                    FilterChip(
+                        selected = prompt == suggestion,
+                        onClick = { prompt = suggestion },
+                        label = { Text(suggestion, style = MaterialTheme.typography.labelSmall) },
+                        shape = CircleShape
+                    )
+                }
+            }
+
+            OutlinedTextField(
+                value = prompt,
+                onValueChange = { prompt = it },
+                modifier = Modifier.fillMaxWidth(),
+                shape = RoundedCornerShape(20.dp),
+                label = { Text(stringResource(R.string.comput_commandium_label)) },
+                placeholder = { Text(stringResource(R.string.comput_commandium_placeholder)) },
+                trailingIcon = {
+                    IconButton(
+                        onClick = { requestCommandium() }
+                    ) {
+                        Icon(
+                            imageVector = Icons.AutoMirrored.Rounded.Send,
+                            contentDescription = stringResource(R.string.comput_command_send)
+                        )
+                    }
+                },
+                keyboardOptions = KeyboardOptions(
+                    imeAction = ImeAction.Send
+                ),
+                keyboardActions = KeyboardActions(
+                    onSend = { requestCommandium() }
+                ),
+                maxLines = 3
+            )
+
+            Button(
+                onClick = { requestCommandium() },
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(48.dp),
+                shape = CircleShape
+            ) {
+                if (isGenerating) {
+                    LoadingIndicator(Modifier.size(18.dp), MaterialTheme.colorScheme.onPrimary)
+                } else {
+                    Text(stringResource(R.string.comput_ask_commandium), fontWeight = FontWeight.Bold)
+                }
+            }
+
+            val outcome = result
+            if (outcome != null) {
+                val outcomeText = outcome.getOrNull()
+                val isError = outcomeText == null
+                val displayText = outcomeText ?: buildString {
+                    append("Error: ")
+                    append(outcome.exceptionOrNull()?.message ?: "Unknown error")
+                }
+                Surface(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .semantics { liveRegion = LiveRegionMode.Polite },
+                    color = if (isError) MaterialTheme.colorScheme.errorContainer else MaterialTheme.colorScheme.surfaceContainerHigh,
+                    shape = RoundedCornerShape(16.dp)
+                ) {
+                    Column(modifier = Modifier.padding(14.dp)) {
+                        Text(
+                            text = stringResource(R.string.comput_generated_command),
+                            style = TextStyle(fontFamily = FontFamily.Monospace, fontSize = 11.sp, fontWeight = FontWeight.Bold),
+                            color = if (isError) MaterialTheme.colorScheme.onErrorContainer else MaterialTheme.colorScheme.primary
+                        )
+                        Spacer(modifier = Modifier.height(4.dp))
+                        SelectionContainer {
+                            Text(
+                                text = displayText,
+                                style = TextStyle(fontFamily = FontFamily.Monospace, fontSize = 13.sp)
+                            )
+                        }
+                        Spacer(modifier = Modifier.height(10.dp))
+                        Row(
+                            modifier = Modifier.fillMaxWidth(),
+                            horizontalArrangement = Arrangement.spacedBy(8.dp)
+                        ) {
+                            Button(
+                                onClick = {
+                                    if (!isError) {
+                                        onUseCommand(outcomeText!!)
+                                        onDismiss()
+                                    }
+                                },
+                                modifier = Modifier.weight(1f),
+                                shape = CircleShape
+                            ) {
+                                Text(stringResource(R.string.comput_use_command))
+                            }
+                            IconButton(
+                                onClick = {
+                                    onCopy(displayText)
+                                }
+                            ) {
+                                Icon(Icons.Rounded.ContentCopy, contentDescription = stringResource(android.R.string.copy))
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        },
+        confirmButton = {},
+
+        dismissButton = {
+            TextButton(onClick = { onDismiss() }) {
+                Text(stringResource(android.R.string.cancel))
+            }
+        },
+        containerColor = MaterialTheme.colorScheme.surface,
+        shape = RoundedCornerShape(28.dp)
+    )
+}
