+++ manager/src/main/java/moe/shizuku/manager/utils/AiExplainUtil.kt
@@ -72,15 +72,6 @@ object AiExplainUtil {
             model = model,
             systemPrompt = null,
             userPrompt = requestPrompt
-        ).map { sanitizeCommand(it) }
+        ).map { CommandSanitizer.sanitize(it) }
     }
-
-    private fun sanitizeCommand(raw: String): String {
-        var t = raw.trim()
-        t = REGEX_THINKING.replace(t, "")
-        return REGEX_FENCE.replace(t) { m -> m.groupValues[1] }.trim()
-    }
-
-    private val REGEX_THINKING = Regex("(?is)<thinking>.*?</thinking>")
-    private val REGEX_FENCE = Regex("(?is)^```[a-zA-Z0-9_-]*\\s*\\n(.*?)\\n```\\s*$")
 }
